+++ src/Reface.App.Starter/AppSetup.cs
@@ -1,5 +1,6 @@
 ﻿using Reface.AppStarter.AppContainerBuilders;
 using Reface.AppStarter.AppContainers;
+using Reface.AppStarter.AppModulePrepairs;
 using Reface.AppStarter.AppModules;
 using Reface.AppStarter.AppSetupPlugins;
 using Reface.AppStarter.AppSetupPlugins.Arguments;