+++ src/Reface.App.Starter/Ext.cs
@@ -47,6 +47,12 @@ namespace Reface.AppStarter
 
         #region Work
 
+        /// <summary>
+        /// 开启一个工作单元
+        /// </summary>
+        /// <param name="app"></param>
+        /// <param name="workName"></param>
+        /// <returns></returns>
         public static IWork BeginWork(this App app, string workName)
         {
             var container = app.GetAppContainer<IComponentContainer>();
@@ -55,6 +61,11 @@ namespace Reface.AppStarter
 
         }
 
+        /// <summary>
+        /// 发布一个事件
+        /// </summary>
+        /// <param name="work"></param>
+        /// <param name="event"></param>
         public static void PublishEvent(this IWork work, Event @event)
         {
             IEventBus eventBus = work.CreateComponent<IEventBus>();